Package name change COMPLETE
