+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.Poli.home"
+        applicationId "edu.cascadia.mobas.pollinatorpathway"
         minSdk 21
         targetSdk 31
         versionCode 1
