+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.Poli.pollinatorpathway">
+    package="edu.cascadia.mobas.pollinatorpathway">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Home">
         <activity
-            android:name="com.Poli.edu.cascadia.mobas.pollinatorpathway.MainActivity"
+            android:name="edu.cascadia.mobas.pollinatorpathway.MainActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
