+++ app/src/main/java/edu/cascadia/mobas/pollinatorpathway/ui/dashboard/DashboardViewModel.java
@@ -1,16 +1,16 @@
-package com.Poli.pollinatorpathway.ui.notifications;
+package edu.cascadia.mobas.pollinatorpathway.ui.dashboard;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class DashboardViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public DashboardViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is dashboard fragment");
     }
 
     public LiveData<String> getText() {
