+++ app/src/main/java/edu/cascadia/mobas/pollinatorpathway/ui/home/HomeViewModel.java
@@ -1,16 +1,16 @@
-package com.Poli.pollinatorpathway.ui.dashboard;
+package edu.cascadia.mobas.pollinatorpathway.ui.home;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class HomeViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public HomeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
