+++ app/src/main/java/edu/cascadia/mobas/pollinatorpathway/ui/notifications/NotificationsViewModel.java
@@ -1,16 +1,16 @@
-package com.Poli.pollinatorpathway.ui.home;
+package edu.cascadia.mobas.pollinatorpathway.ui.notifications;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class NotificationsViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public NotificationsViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is notifications fragment");
     }
 
     public LiveData<String> getText() {
