+++ app/src/test/java/edu/cascadia/mobas/pollinatorpathway/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.Poli.pollinatorpathway;
+package edu.cascadia.mobas.pollinatorpathway;
 
 import org.junit.Test;
